+++ desktop/packages/mullvad-vpn/src/main/grpc-type-convertions.ts
@@ -17,6 +17,7 @@ import {
   DeviceEvent,
   DeviceState,
   DirectMethod,
+  DomainFronting,
   EncryptedDnsProxy,
   EndpointObfuscationType,
   ErrorStateCause,
@@ -1164,11 +1165,11 @@ function fillApiAccessMethodSetting<T extends grpcTypes.NewAccessMethodSetting>(
       accessMethod.setEncryptedDnsProxy(encryptedDnsProxy);
       break;
     }
-    // case 'domain-fronting': {
-    //   const domainFronting = new grpcTypes.AccessMethod.DomainFronting();
-    //   accessMethod.setDomainFronting(domainFronting);
-    //   break;
-    // }
+    case 'domain-fronting': {
+      const domainFronting = new grpcTypes.AccessMethod.DomainFronting();
+      accessMethod.setDomainFronting(domainFronting);
+      break;
+    }
     default:
       accessMethod.setCustom(convertToCustomProxy(method));
   }
@@ -1265,9 +1266,9 @@ function convertFromApiAccessMethodSettings(
       "no 'Encrypted DNS proxy' access method was found",
     ),
   ) as AccessMethodSetting<EncryptedDnsProxy>;
-  // const domainFronting = convertFromApiAccessMethodSetting(
-  //   ensureExists(accessMethods.getDomainFronting(), "no 'Domain fronting' access method was found"),
-  // ) as AccessMethodSetting<DomainFronting>;
+  const domainFronting = convertFromApiAccessMethodSetting(
+    ensureExists(accessMethods.getDomainFronting(), "no 'Domain fronting' access method was found"),
+  ) as AccessMethodSetting<DomainFronting>;
   const custom = accessMethods
     .getCustomList()
     .filter((setting) => setting.hasId() && setting.hasAccessMethod())
@@ -1279,7 +1280,7 @@ function convertFromApiAccessMethodSettings(
     direct,
     mullvadBridges: bridges,
     encryptedDnsProxy,
-    // domainFronting,
+    domainFronting,
     custom,
   };
 }
@@ -1316,8 +1317,8 @@ function convertFromAccessMethod(method: grpcTypes.AccessMethod): AccessMethod {
       return { type: 'bridges' };
     case grpcTypes.AccessMethod.AccessMethodCase.ENCRYPTED_DNS_PROXY:
       return { type: 'encrypted-dns-proxy' };
-    // case grpcTypes.AccessMethod.AccessMethodCase.DOMAIN_FRONTING:
-    //   return { type: 'domain-fronting' };
+    case grpcTypes.AccessMethod.AccessMethodCase.DOMAIN_FRONTING:
+      return { type: 'domain-fronting' };
     case grpcTypes.AccessMethod.AccessMethodCase.CUSTOM: {
       return convertFromCustomProxy(method.getCustom()!);
     }