+++ ios/MullvadVPN/Coordinators/Settings/APIAccess/Models/AccessMethodViewModel+Persistent.swift
@@ -66,6 +66,8 @@ extension AccessMethodViewModel {
             .bridges
         case .encryptedDNS:
             .encryptedDNS
+        case .domainFronting:
+            .domainFronting
         case .socks5:
             try socks.intoPersistentProxyConfiguration()
         case .shadowsocks: